The assembler is now mostly complete
Finally got all instructions (bar TRAP) working, all Directives working,
and Labels working. Immediate values should also work completely -- also
added that if a word is all digits, then it's just a decimal number.

All that needs to be done is outputting the assembled code to a file,
as well as optimise the code, and some documentation.

There's also some minor fixes to create better error's/warning's for the
user.