Trap instruction implemented
The TRAP instruction is now working.

.BLKW now accepts a 2nd argument that can come in the form of either
an immediate value or a label. If it's an immediate value, the blocks
are initialised to the value of the immediate. If it's a label (and is
a valid label), the address of the label is the initialiser.

A new invalid_argument_count error was created to tell the user how
many arguments were supplied vs how many were expected.