Added two features
There is now a NEG instruction. This will create two instructions,
the first creating the 1's compliment of the provided register, the
second actually making it negated. In the future, this should be
expanded to allow multiple registers -- one for storing the result.
        i.e. NEG R2, R1 should store the negated version of R1 into R2
As it stands, this is the NEG instruction:
        NEG R1
This is what it produces:
        NOT R1, R1
        ADD R1, R1, #1

There is now a SUB instruction. This will first generate a NEG instruction,
then add it to the supplied register. In the future this should be expanded
to allow more registers.
As it stands, this is the SUB instruction:
        SUB R2, R1
This is what it produces:
        NEG R1
        ADD R2, R2, R1
        NEG R1
While the original value of R1 is not lost in this scenario, the way this
is implemented can cause problems, e.g.:
        SUB R1, R1
Leads to the following being generated:
        NEG R1         ; R1 =  1  After, R1 = -1
        ADD R1, R1, R1 ; R1 = -1  After, R1 = -2
        NEG R1         ; R1 = -2  After, R1 =  2
This would result in R1 being set to the incorrect value (in this case, 2
instead of 0).