Some bug fixes
Fixed a problem where the assembler would throw an error if the line
ended with a space character.

Fixed a problem where the assembler would think two instructions have
the same address if the first one has instructions on the same line.

Fixed a problem where assembling a line starting with a label containing
an instruction, which itself referenced a label not yet defined, it would
throw an error.