A few changes
Disassembly is now done by individual tokens.

The lst file is now better written to instead of just generating random
disassembled output, it now is aware of where in the original file it is.

Added a new directive: .SET
.SET takes a register and an immediate value (might change this later to
allow a label). If the immediate value is too large to use in a single
ADD instruction, it will generate 3 instructions -- a BR to jump over the
next instruction, which is a .FILL directive that fills the memory at that
address with the value, and then a LD instruction which will put the
value at the address before it in memory into the provided register.
If the value is able to be put into the register with a simple ADD
instruction, then the register is first cleared before then adding
the value to the register.