Symbol matching for errors
If a symbol isn't found it is now checked against each symbol in the
symbol table, using the Levenshtein distance algorithm. If there is
a close enough match, we will also print out the match when telling
the user that the symbol doesn't exist.

Some more logic checks have been added.