Added new directive: .LSHIFT
.LSHIFT takes a register, and a 4 bit unsigned immediate value. With
this value it left shifts the value in the register by that amount.
Basically, it just generates that many ADD's.

Fixed: A problem with calculating .SUB's memory requirement. It would
       guess 5, but really meant 3.

Fixed: If there were 2 or more .ORIG directions, the assembler would
       reset the internal program counter to the provided offset.

Improved: Error's when the address/immediate value provided to an
          instruction which was out of range.