Basic Include functionality
A new directive .INCLUDE has been added. This directive
takes a string providing a file name as an operand. If this
file is found and can be opened, the lexer will go through
this file as well, and put all tokens found in the same space.

From the rest of the first-passes perspective, it will just
be one large file (as with the second pass).

It's basic, as I'm still trying to work out how the LST file
should work. It's possible to keep backwards functionality by
wrapping certain things in #ifdef's, however I haven't really
worked out how I want the new LST file to look like. Should it
also contain file names?