Diagnostic added, CMake improved
Added an invalid argument diagnostic. It's a little problematic
in that, if the last argument provided is a string or a character
literal, it will highlight one less character than it should, e.g.:

.STRINGZ "Here" "is a weird string"
         ~~~~~~~~~~~~~~~~~~~~~~~~~

I've moved any files relevant to addons into conditional
compilation with CMake, as the way it was done before (ifdef around
the entire class) was bad, and I felt bad about it.