Command line parser, new example, minor changes
Fixed some issues that gcc had when compiling (due to the name of
the editor being the same as the class name).

A command line parser now allows the user to specify a file to open
on startup.

There's a new example Examples/PRNG.asm, which is a rather horrid
implementation of a simple PRNG algorithm that I hope to improve at some
point to a level of adding some of the subroutines as trap routines.