Minor logic fix
Before now, if a logic error was placed in such a way that a label
was in front of the instruction, no warning would be placed, i.e.:
TEST  BRn #0

The above wouldn't throw a warning about a superfluous statement
because of the way I checked for problems.

This commit should fix this, as well as add a warning for assembling
RTI instructions.