Performance improvements, bug fixes
Performance Improvment:
Removed all regex's as they were slow, and replaced them with easier
checks (this effects how immediate values are found, and how labels
are identified).

Bug fix:
Fixed a bug where if the assembly file had a single '#' it was treated
as a valid decimal value (this was introduced with the above performance
improvement).

I'm hoping to improve (both performance, and bug wise) the new
tokeniser at some point.