Switched to LLVM style
Due to a Visual Studio Code update, there were glitches that caused the
style to change repeatedly, causing me to remove some plugins, and turn
to the LLVM styling (which I thought I already was using, but apparently
not).

There are also a couple of changes as well, like expanding use of C++-17
(e.g. Structured bindings and string_view).